+++ common/src/main/java/net/mehvahdjukaar/amendments/AmendmentsClient.java
@@ -96,8 +96,6 @@ public class AmendmentsClient {
         ClientHelper.registerRenderType(ModRegistry.HANGING_FLOWER_POT.get(), RenderType.cutout());
         ClientHelper.registerRenderType(ModRegistry.WALL_LANTERN.get(), RenderType.cutout());
         MenuScreens.register(ModRegistry.LECTERN_EDIT_MENU.get(), LecternBookEditScreen::new);
-
-
     }
 
     public static void lateClientSetup() {