+++ common/src/main/java/net/mehvahdjukaar/amendments/common/tile/LiquidCauldronBlockTile.java
@@ -29,7 +29,10 @@ import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import org.jetbrains.annotations.NotNull;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
 
 public class LiquidCauldronBlockTile extends BlockEntity implements IExtraModelDataProvider, ISoftFluidTankProvider {
@@ -64,6 +67,8 @@ public class LiquidCauldronBlockTile extends BlockEntity implements IExtraModelD
         List<MobEffectInstance> combinedEffects = new ArrayList<>();
         List<MobEffectInstance> existingEffects = PotionUtils.getAllEffects(tankTag);
         List<MobEffectInstance> newEffects = PotionUtils.getAllEffects(newFluid.getTag());
+        if (newEffects.equals(existingEffects)) return;
+
         float oldMult = oldCount / (float) newCount;
         float newMult = 1 - oldMult;
         combineEffects(combinedEffects, existingEffects, oldMult);
@@ -84,8 +89,8 @@ public class LiquidCauldronBlockTile extends BlockEntity implements IExtraModelD
 
     @NotNull
     private static MobEffectInstance mergeEffects(MobEffectInstance e, MobEffectInstance e1) {
-        return new MobEffectInstance(e.getEffect(), (int) ((e.getDuration() + e1.getDuration()) / 2f),
-                (int) ((e.getAmplifier() + e1.getAmplifier()) / 2f));
+        return new MobEffectInstance(e.getEffect(), (e.getDuration() + e1.getDuration()),
+                (e.getAmplifier() + e1.getAmplifier()));
     }
 
     public static void saveEffects(CompoundTag tag, Collection<MobEffectInstance> effects) {
@@ -100,6 +105,7 @@ public class LiquidCauldronBlockTile extends BlockEntity implements IExtraModelD
                                        List<MobEffectInstance> current, float mult) {
         for (var e : current) {
             MobEffect effect = e.getEffect();
+
             MobEffectInstance newInstance;
             if (effect.isInstantenous()) {
                 newInstance = new MobEffectInstance(effect, e.getDuration(), (int) (e.getAmplifier() * mult));
@@ -120,12 +126,11 @@ public class LiquidCauldronBlockTile extends BlockEntity implements IExtraModelD
         int oldAmount = tankFluid.getCount();
         int newAmount = newFluid.getCount();
         CompoundTag combinedTag = new CompoundTag();
-        combinedTag.putInt(DyeBottleItem.COLOR_TAG, new RGBColor(oldColor).asHCL()
-                .mixWith(new RGBColor(newColor).asHCL(), (float) newAmount / (oldAmount + newAmount))
-                .asRGB().toInt());
+        combinedTag.putInt(DyeBottleItem.COLOR_TAG, DyeBottleItem. mixColor(oldColor, newColor, oldAmount, newAmount));
 
         tankFluid.setTag(combinedTag);
     }
+
 
 
 
@@ -179,7 +184,7 @@ public class LiquidCauldronBlockTile extends BlockEntity implements IExtraModelD
 
         BlockState state = this.getBlockState();
 
-        if(state.getBlock() instanceof ModCauldronBlock cb){
+        if (state.getBlock() instanceof ModCauldronBlock cb) {
             state = cb.updateStateOnFluidChange(state, fluidTank.getFluid());
         }
 
@@ -197,7 +202,7 @@ public class LiquidCauldronBlockTile extends BlockEntity implements IExtraModelD
     public boolean handleInteraction(Player player, InteractionHand hand) {
         //interact with fluid holder
         if (this.fluidTank.interactWithPlayer(player, hand, level, worldPosition)) {
-             this.setChanged();
+            this.setChanged();
             return true;
         }
         return false;